chromeos-setimage: simplify verified rootfs detection
Instead of making guesses based on the booted system,
chromeos-setimage will now check the target kernel to see if
root=/dev/dm-0 or not.  If it is, then the kernel is verified
and we default to installing a verified rootfs.  A flag is added
for legacy/efi-supporting bootloaders to allow the command-line
flags to override the kernel partition values.

TEST=tested recovery, regular withdev, regular nowithdev, factory still underway
BUG=chromium-os:6662

Change-Id: Ib00173a236a5fc94d01d1dd7ff9cfbcd96448c45

Review URL: http://codereview.chromium.org/3418011