Increase the number of tries for autoupdated image.
In certain cases we may need to reboot a couple of times following an
autoupdate, due to BIOS and/or EC changes. We don't want that to cause
unnecessary failures.

Change-Id: Ieb042fc0b70fe7cc3d2b384425484565d7a6de92

BUG=chromium-os:8249
TEST=manual

Install the image containing this fix, then trigger an autoupdate. The
postinst script on the current image should set the number of tries on the
new image to 6 instead of the usual 3. After the autoupdate has finished,
but before the new image has been booted, obtain a root shell and run "cgpt
show /dev/sda". You should see the number of tries for the new kernel
partition set to 6. If you boot the new image and wait more than 30 seconds
or so, the tries will be set to zero and the successful flag set to one, so
you won't see any difference.

Review URL: http://codereview.chromium.org/4098006