Leave priority unchanged if kernel subkey changes on try_b firmware updates.
When we upgrade the firmware, we would normally write to firmware B and set
the CMOS flag to try firmware B on the next boot. That doesn't always work
(chrome-os-partner:1563), so we may sometimes find ourselves booting
firmware A instead of B. We've already put in a workaround in the shellball
to quickly reboot into firmware B if that happens
(http://codereview.chromium.org/4339002/), but we need another workaround to
handle yet another special case.

This change handles the case where the firmware update involves changing the
kernel subkey. If we do that, firmware B can't boot the old kernel and
firmware A can't boot the new kernel. If the next boot is to firmware B, all
is good. If it's to firmware A, firmware A will invalidate the new kernel
before trying the old kernel, which undoes our update.

The workaround is that if the firmware update changes the kernel subkey,
then we leave the kernel priorities unchanged, so that if we boot into
firmware A, it will try the old kernel first without invalidating the new
kernel. The old kernel/rootfs will then reboot into firmware B, which can
successfully boot the new kernel.

After the new kernel boots, we want to modify the priority so that the new
kernel will be booted first, to prevent always trying the old kernel (and
now invalid) kernel first.

Got all that? Yeah, it's a bit of a hack.

Anyhoo, this seems to work.

BUG=chrome-os-partner:1654
TEST=manual

Testing is a pain, because the problem only turns up when:

1. We're autoupdating (not a manual install or recovery), AND
2. We're upgrading the firmware politely (ie, only firmware B, with no forced
   reboot), AND
3. Firmware B has a differerent kernel subkey from firmware A, AND
4. The device under test loses its CMOS state during the next reboot
   (chrome-os-partner:1563), which typically means it's a Mario running on
   battery alone and is power cycled after the AU.

However, this fix *should* work even if it's only in the image that we're
autoupdating TO. That is, you should be able to start from an image that
doesn't have this change in it, and as long as the new image does have the
change, it should work. Of course, if you start with too old an image it may
not.

[Sorry, no, I don't know how old is too old, but I *think* that any image
that actually runs the postinst script from the newly upgraded rootfs
instead of the old rootfs will work. Note that the old image needs to
have the fix for chrome-os-partner:1563 or you'll have other problems.]

Setting up the AU server and device under test isn't hard. The tricky bit is
generating an image and firmware shellball containing a different kernel
subkey. Because we're changing keys, we'll need the device under test to be
signed with devkeys, not secret keys. Here's how to generate a new set of
keys (all this in the chroot, of course):

* Copy the existing devkeys somewhere where we can fiddle with them:

  mkdir /tmp/foo
  cp -r .../src/vboot_reference/tests/devkeys /tmp/foo/

* Make a copy of that, to fiddle with:

  cd /tmp/foo
  cp -r devkeys newkeys

* In the newkeys directory, edit create_new_keys.sh so to comment out six
  of the seven lines that create keys, so that only the kernel subkey will
  be affected.

  cd newkeys
  vi ./create_new_keys.sh

  The relevant section should look like this:

    # Create the normal keypairs
    #NO make_pair root_key                 11
    #NO make_pair firmware_data_key        7
    make_pair kernel_subkey            7
    #NO make_pair kernel_data_key          4

    # Create the recovery and factory installer keypairs
    #NO make_pair recovery_key             11
    #NO make_pair recovery_kernel_data_key 11
    #NO make_pair installer_kernel_data_key 11

* Now run that modified script, in that directory.

  ./create_new_keys.sh

  If you compare the devkeys directory with the newkeys directory, these
  files should be different:

    kernel.keyblock
    kernel_subkey.vbprivk
    kernel_subkey.vbpubk

Now we're ready to produce the AU image.

* Grab the latest buildbot chromiumos_base_image.bin for containing this
  change. You want the raw Mario image and not the recovery image, of
  course.

  For this example, I put it in /tmp/foo/

* Change the version strings in /etc/lsb-release, so that you can tell if it
  updated:

  cd src/scripts
  ./mount_gpt_image.sh -f /tmp/foo -i chromiumos_base_image.bin
  sudo vi /tmp/m/etc/lsb-release
  ./mount_gpt_image.sh -u

* Tag the image so that it will do a firmware update:

  cd ../platform/vboot_reference/scripts/image_signing/
  ./tag_image.sh --from /tmp/foo/chromiumos_base_image.bin --update_firmware 1

* Set the password to something you know, so you can log in after the update
  is finished and run flashrom or an old shellball to restore your device
  back to the devkeys:

  ./set_chronos_password.sh /tmp/foo/chromiumos_base_image.bin chronos

* Sign the modified image with the keys that we produced earlier:

  sudo ./sign_official_build.sh ssd /tmp/foo/chromiumos_base_image.bin \
    /tmp/foo/newkeys /tmp/foo/new_image.bin

* Now we can use that newly signed image to produce the AU binary:

  cd .../src/scripts
  sudo ./cros_generate_update_payload --image /tmp/bar/new_image.bin \
     --noold_style --output update.gz

Use that new AU binary to test the upgrade. After the AU has been applied,
unplug and power the machine off instead of just rebooting. When you reboot,
it should boot into firmware A, reboot into firmware B, launch the new
kernel, copy firmware B into firmware A, and eventually (after a minute or
two when the autoupdate engine has had a chance to run once), mark the new
kernel partition as the highest priority.

Remember that if all that works, you'll have modified your firmware keys so that the standard devkey-signed images won't boot anymore. You'll have to reflash the BIOS (manually or via recovery mode) to put your machine back to normal.

Review URL: http://codereview.chromium.org/5381001

Change-Id: I6aa360995437a6b0a5e8b57089232aef0e261772