chromeos-postinst: force to re-generate VPD cached log file if activate_date exists.
There is a race condition in init/chromeos_startup script, and comes out a bug
cannot be fixed in time because the factory image has been released. We use this
patch in AU to solve that bug.

This patch should be run only for:
  1. the specific model (which /usr/sbin/activate_date exists on)
  2. if the current image is before R12, assuming the bug is fixed in R12.

please read chrome-os-partner:3412 for more details.

Change-Id: I185e996e0771e42d297558fa8418ca7eb3be622d

R=adlr@chromium.org
BUG=chrome-os-partner:3412
TEST=

Review URL: http://codereview.chromium.org/6873146