Fix losetup race conditions in install scripts.
The following pattern creates a race condition:

   LOOP_DEV=$(sudo losetup -f)
   ...
   sudo losetup "${LOOP_DEV}" "${ROOT_FS_IMG}"

If two steps similar to the above run in parallel, and both steps pick up
the same free loop device, they may try to mount different images with the
same loop device and one of the two stages will get a "device is busy" error.

To fix this, we should switch to the following pattern:
   LOOP_DEV=$(sudo losetup --show -f "${ROOT_FS_IMG}")

This CL implements the above.

BUG=chromium-os:18046
TEST=Run image_to_usb.sh --install to install image to USB stick (and
     test the installer script.)

Change-Id: I59dbdc9fb75978c11baf0bcbf19a1ad69d886c02
Reviewed-on: http://gerrit.chromium.org/gerrit/4588
Reviewed-by: Darin Petkov <petkov@chromium.org>
Tested-by: David James <davidjames@chromium.org>