Restore the --run_as_root option pending a better solution.
The developer mode installation for Cr-48 uses an old recovery
kernel that still passes the --run_as_root option; until that kernel
can be replaced, --run_as_root has to be accepted and ignored.  The
troublesome recovery kernel itself can't be removed or updated right
away because of the need for short-term support of old installers
that still require --run_as_root.

BUG=chromium-os:20684
TEST=install an image with the change following the online instructions

Change-Id: Ic2d9ef2ab34a959a2b42736846bba31d7725b171
Reviewed-on: http://gerrit.chromium.org/gerrit/8142
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>