Create initial C++ conversion of chromeos-chroot-postinst.
This is the first pass of a static postinstall script to support the 32 to 64
bit transition. This is DISABLED, but should be ready for use, but with these
limitations:

1) Breaks installs on non-Chrome hardware
2) Shells out for verity
3) Shells out for cgpt
4) Unittests are linked into main executable
5) Extra library dependancies (libbase, especially)

All of these problems should be addressed in follow on CLs over the next few
days.

To ENABLE uncomment two lines at the beginning of chromeos-postinst.

BUG=chromium-os:25372
TEST=x86-mario live updates, 64 bit testing, and USB install

Change-Id: I40d27dcef36720febc4d03e4be785e438a961431
Reviewed-on: https://gerrit.chromium.org/gerrit/15026
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Commit-Ready: Don Garrett <dgarrett@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>