Re-enable Legacy Bios Support
This change adds code to postinstall to correctly update configuration files
for the syslinux boot loader, if the current kernel was booted via syslinux. A
skeleton handler for EFI is present, but it will fail.

The --legacy command line option is removed from cros_installer, and the
chromeos-postinst script no longer tries to use it when it was needed.

The Partition class gained a uuid lookup method.

Overall postinst logic was moved from cros_installer_main.cc to
chromeos_postinst.cc.

inst_utils gained the helpers CopyFile, and SetKernelArg (a match for
ExtractKernelArg).

BUG=chromium-os:25395
TEST=Extended/ran unittests, tested install/image_to_live to an EEPC.

Change-Id: Id6b2ba861ac5301e61138915155a37feedf22d6c
Reviewed-on: https://gerrit.chromium.org/gerrit/21129
Tested-by: Don Garrett <dgarrett@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Ready: Don Garrett <dgarrett@chromium.org>