install_gpt: move rootfs to beginning of disk on rotating media
This adds a flag and detection to to the install_gpt function
which provides an alternative disk layout for machines which are
using a rotating disk.  We move the rootfs partitions to be before
the stateful partition in order to get better sequential disk
performance.

BUG=chrome-os-partner:12203
TEST=ad hoc: install an image via chromeos-install on a system with
 a rotating disk, inspect partition layout, rootfs should be before
 stateful
TEST=ad hoc: install an image via chromeos-install on a system with
 a non-rotating disk, inspect partition layout, layout should be
 identical to an install from before this change
TEST=`cbuildbot amd64-generic-full`

Change-Id: I637de54be9deba18dc75cbe2a476d999dc83576c
Reviewed-on: https://gerrit.chromium.org/gerrit/29726
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Sonny Rao <sonnyrao@chromium.org>
Commit-Ready: Sonny Rao <sonnyrao@chromium.org>