Postinst removes /var/lib/preload-network-drivers.
A cache file created on stateful to speed loading of wifi drivers is sometimes
incorrect after a reboot which will prevent the correct wifi driver from
loading. This change deletes the file in question during postinst to avoid
the corruption problem.

This will slow the next boot slightly, but should be safe since the cache
file is simply recreated during boot if not present.

BUG=chrome-os-partner:13259
TEST=image_to_live, verify file missing, reboot, verify it returned.

Change-Id: Ic4aa1c162944a72d9e7ee4283035bfc331e05c31
Reviewed-on: https://gerrit.chromium.org/gerrit/31772
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Ready: Don Garrett <dgarrett@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>