remove(chromeos-install): Unused and certainly not correct.
This is the one script that called /usr/sbin/write_gpt.sh which is
generated when build_image runs. However due to our extensive use of
image_to_vm.sh the final partition table for a typical install is chosen
after the rootfs is generated so write_gpt.sh will be wrong.

Since we don't have the use case that chromeos-install filled in the
first place this can just go away. Eventually the rest of this should go
too, it is almost all dead code. The little bit of "legacy" postinst
code we need can probably move to cgpt and a bash script. Someday...