fix(coreos-setgoodroot): Remove invalid use of 'local'.
Turns out this error wasn't harmless after all because -e was set in the
shebang line. This means our system never marks newly upgraded systems
as good and will roll back after the next reboot. This appears to be a
bug that has existed since the script's introduction in a51e35b2 so it
is really surprising it hasn't come up before.