cleanup(chromeos-postinst): Remove support for alternate BIOS types.
Even when we run on a BIOS that needs no bootloader or special handling
we should still do the same updates in the boot partition. Otherwise
doing things like switching a host between UEFI and legacy mode may lead
to some unexpected results. With this the cros_legacy flag also dies.

Nothing really changes for us, this was pretty much all dead or broken
code.